Moved back various definitions from KeyDef.hpp to Handler.hpp
