+++ deps/first/Input/include/Input/KeyDef.hpp
@@ -3,7 +3,6 @@
 #include <stdexcept>
 #include <cstdint>
 #include <initializer_list>
-#include <functional>
 
 namespace RC::Input
 {
@@ -304,27 +303,4 @@ namespace RC::Input
     auto operator&(const ModifierKeys& keys, const ModifierKeys& key) -> bool;
 
     auto operator++(Input::Key& key) -> Input::Key&;
-
-    using EventCallbackCallable = std::function<void()>;
-
-    struct InputEvent
-    {
-        Key key;
-        ModifierKeys modifier_keys{};
-    };
-
-    struct KeyData
-    {
-        ModifierKeys required_modifier_keys{};
-        EventCallbackCallable callback{};
-        uint8_t custom_data{};
-        void* custom_data2{};
-        bool requires_modifier_keys{};
-        bool is_down{};
-    };
-
-    struct KeySet
-    {
-        std::unordered_map<Key, std::vector<KeyData>> key_data;
-    };
 } // namespace RC::Input
